Add files via upload
This Python script allows users to trace the location of a phone number by using the phonenumbers library to identify the country and mobile carrier. It then uses the OpenCage Geocoding API to retrieve geographic coordinates (latitude and longitude) of the identified location. Finally, the folium library is used to mark the location on a map, which is saved as an HTML file for visualization.